refactor: relocate firebase config and guard analytics
